Added UI screens
Added a start screen, game over screen and victory screen. The screens are designed in a way that allows easy creation of new ones by copy pasting an existing scene and swapping out audio / visuals / texts as needed. Screens feature a blinking text effect, background music, transition to a new screen upon pressing a button that plays a sound and a transition animation